Few more fixes. Tried speeding up portfolio get list. Fixed autocomplate dropdown disapearing behind modal

+++ Website/Models/DTOs/AddressDTO.cs
@@ -23,7 +23,7 @@ namespace Website.Models.DTOs
 
         public string Town { get; set; }
         public string City { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
